LavaHack b0.1.6.5-9 | Changes #117: Client:     Modules:         Added AutoAnchor         Added AirPlace(WIP, probably does not work)         Added Gui -> Shaders(group) (WIP, no fps xD)         Added PvPResources -> Style -> Horizontal(HUD)         Moved SilentXP from COMBAT to PLAYER         Moved PlayerLook from EXPLOIT to WIP(RENDER) (it may give crash)         Moved SoftScaffold from MOVEMENT to WIP         Improved Octopus         Cleanupped PacketEat         Rewrote placement of FlattenRewrite         Removed AutoClicker         Removed Surround(goofy ah module tbh)
    Command:
        Added fakeplayer command

    Click Gui:
        Added shader feature

    Stuff:
        Added extended block system
        Added InventoryUtil::findBlockExtended
        Added InventoryUtil::findBlockExtendedExclude
        Added ShaderHelper
        Cleanupped InventoryUtil
        Cleanupped EntityUtil
        Cleanupped MovementUtil
        Rewrote WorldUtilKt::entityPosition
        Removed RotationProvider

        Patterns:
            Added shader pattern
            Added placement pattern

        Extended Block System:
            Added respawn anchor

+++ src/main/java/com/kisman/cc/util/world/WorldUtil.kt
@@ -41,7 +41,7 @@ fun updateCamera(camera : ICamera, entity : Entity) {
     camera.setPosition(entity.posX, entity.posY, entity.posZ)
 }
 
-fun entityPosition(entity : Entity) : BlockPos = BlockPos(entity.posX, entity.posY, entity.posZ)
+fun entityPosition(entity : Entity) : BlockPos = BlockPos(entity.posX.toInt(), entity.posY.toInt(), entity.posZ.toInt())
 
 fun playerPosition() : BlockPos = entityPosition(mc.player)
